requirements.txt
certifi==2017.4.17
chardet==3.0.4
Django==1.11.3
django-cors-middleware==1.3.1
django-extensions==1.8.1
django-mongodb-engine==0.6.0
django-oauth-toolkit==1.0.0
djangorestframework==3.6.3
djangotoolbox==1.8.0
enum34==1.1.6
idna==2.5
mongoengine==0.15.0
mysqlclient==1.3.12
oauthlib==2.0.2
pymongo==3.6.1
pytz==2017.2
requests==2.18.1
six==1.10.0
urllib3==1.21.1
validate-email==1.3
